# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Deliver Swagger APIs",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/v1/orders": {
            "get": {
                "tags": [
                    "Admin > Order"
                ],
                "summary": "Show Orders List",
                "parameters": [
                    {
                        "name": "created_from",
                        "in": "query",
                        "description": "created_from",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "description": "created_to",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "description": "state_id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "status_id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin > Order"
                ],
                "summary": "Store An Order",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "example": {
                                    "code": "123456",
                                    "parcel_value": 20000000,
                                    "type": "on_demand",
                                    "schedule_id": 1,
                                    "cod_amount": 100000000,
                                    "package_quantity": 2,
                                    "priority": "high",
                                    "items": [
                                        {
                                            "material_code": "654321",
                                            "name": "پنیر لبنه 400 گرمی",
                                            "quantity": 2,
                                            "size": "10,20,30",
                                            "weight": 20
                                        }
                                    ],
                                    "locations": [
                                        {
                                            "latitude": 35.721221,
                                            "longitude": 51.37104,
                                            "name": "فروشگاه اپادانا",
                                            "address": "تهران - خیابان خرمشهر - نرسیده به دشتک - پلاک 52 - طبقه 2",
                                            "phone": "09123654789",
                                            "postal_code": "1234568790",
                                            "type": "pick_up"
                                        },
                                        {
                                            "latitude": 35.73122,
                                            "longitude": 51.38104,
                                            "name": "ریحانه حسینی",
                                            "address": "تهران - سعادت اباد - بالاتر از میدان کاج - کوچه ی 7",
                                            "phone": "09123654789",
                                            "postal_code": "1234568790",
                                            "type": "drop_off"
                                        }
                                    ],
                                    "permissions": {
                                        "return": true,
                                        "partial_return": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/v1/order-states": {
            "get": {
                "tags": [
                    "Admin > Order State"
                ],
                "summary": "List all Order Statees",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin > Order State"
                ],
                "summary": "Create a Order State",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "group_id",
                                    "title",
                                    "message"
                                ],
                                "example": {
                                    "title": "title",
                                    "name": "name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/v1/order-statuses": {
            "get": {
                "tags": [
                    "Admin > Order Status"
                ],
                "summary": "List all Order Statuses",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin > Order Status"
                ],
                "summary": "Create a Order Status",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "group_id",
                                    "title",
                                    "message"
                                ],
                                "example": {
                                    "state_id": 1,
                                    "title": "title",
                                    "name": "name",
                                    "sort": 5,
                                    "code": "501",
                                    "next_status_id": 5
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}